requirements.txt
absl-py==1.3.0
asgiref==3.5.2
astunparse==1.6.3
boto3==1.26.16
botocore==1.29.16
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cmake==3.25.0
colorama==0.4.6
crispy-bootstrap5==0.7
cryptography==37.0.4
dj-database-url==0.5.0
Django==4.0
django-countdowntimer-model==0.0.7
django-crispy-forms==1.14.0
django-environ==0.9.0
django-heroku==0.3.1
django-rest-knox==4.2.0
django-ses==3.2.2
django-utils-six==2.0
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
dlib==19.24.0
environ==1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==22.12.6
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.7.0
idna==3.3
jmespath==1.0.1
keras==2.11.0
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==22.0
Pillow==9.2.0
protobuf==3.19.6
psycopg2==2.9.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2020.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==4.0.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
twilio==7.15.0
typing_extensions==4.4.0
tzdata==2022.1
urllib3==1.26.11
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
